requirements.txt
Django==5.1.6
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
python-dotenv==1.0.1
django-filter==24.3
drf-spectacular==0.28.0
Pillow==11.1.0
pandas==2.2.3
openpyxl==3.1.5
scikit-learn==1.6.1
faker==33.3.1
django-redis==5.4.0
